+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Bump Sentry Gradle Plugin to v3.2.1
 - Move splash screen lottie config to XML
 - Show notification permission rationale dialog when user doesn’t allow notification permission
+- Update app icons and strings for Tamil Nadu app
 
 ## 2022-10-31-8480